deployed-files/resume-app/backend/requirements.txt
fastapi==0.95.2
uvicorn==0.21.0
pydantic==1.10.2
torch==1.12.1
transformers==4.38.2
onnxruntime==1.12.0
numpy==1.23.4
llama-cpp-python==0.1.37
scikit-learn==1.0.2
